User can delete own question/answer

+++ app/controllers/questions_controller.rb
@@ -35,8 +35,10 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question successfully deleted.'
+    end
   end
 
   private
